pyproject.toml
[tool.poetry]
name = "boonli-api"
version = "0.1.0"
description = "API allowing you to fecth menus from boonli.com"
authors = ["Allan Beaufour <allan@beaufour.dk>"]
readme = "README.md"
license = "Apache-2.0"
packages = [{ include = "boonli_api" }]
repository = "https://github.com/beaufour/boonli_api"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.1"
bs4 = "^0.0.1"
lxml = "^4.9.1"
requests-toolbelt = "^0.9.1"
python-dateutil = "^2.8.2"
icalendar = "^4.1.0"
functions-framework = "^3.2.0"
typing-extensions = { version = "~4.3.0", python = "<3.8" }

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^22.8.0"
pre-commit = "^2.20.0"
mypy = "^0.971"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
